# app/Http/Resources/v1/RoleResource.php
<?php

namespace App\Http\Resources\v1;

use Illuminate\Http\Resources\Json\JsonResource;

/**
 * @OA\Schema(
 *     title="Role Resource",
 *     description="Role resource",
 *     @OA\Xml(
 *         name="RoleResource"
 *     )
 * )
 */
class RoleResource extends JsonResource
{

    /**
     * @OA\Property(
     *     title="Data",
     *     description="Data wrapper"
     * )
     *
     * @var \App\Models\Role[]
     */
    public function toArray($request)
    {
        return [
            'Role ID' => $this->role_id,
            'Role' => $this->name,
        ];
    }
}