# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Laravel OpenApi User Management Apis Documentation",
        "description": "L5 Swagger OpenApi description",
        "contact": {
            "email": "usamauddin217@gmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:8000/api/v1",
            "description": "Demo API Server"
        }
    ],
    "paths": {
...
        "/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get list of Roles",
                "description": "Returns list of Roles",
                "operationId": "get roles",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleResource"
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Store Single new Role",
                "description": "Returns Inserted Role Resource",
                "operationId": "store role",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
        "/roles/{id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get Role information",
                "description": "Returns Role Resource",
                "operationId": "getRoleById",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            },
            "put": {
                "tags": [
                    "Roles"
                ],
                "summary": "Update existing Role",
                "description": "Returns updated Role Resource",
                "operationId": "updateRole",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "Role id",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/RoleRequest"
                            }
                        }
                    }
                },
                "responses": {
                    "202": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/RoleResource"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    },
                    "404": {
                        "description": "Resource Not Found"
                    }
                }
            }
        },
        "/roles/permission": {
            "post": {
                "tags": [
                    "RolePermissions"
                ],
                "summary": "Store new Role Permission",
                "description": "Returns Role permission Resource",
                "operationId": "storeRolePermisson",
                "requestBody": {
                    "required": true
                },
                "responses": {
                    "201": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "Role ID": {
                                            "type": "number",
                                            "example": 1
                                        },
                                        "Role Name": {
                                            "type": "string",
                                            "example": "Hr Manager"
                                        },
                                        "Permission ID": {
                                            "type": "string",
                                            "example": "2"
                                        },
                                        "Permission Name": {
                                            "type": "string",
                                            "example": "Update Employee Information"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "401": {
                        "description": "Unauthenticated"
                    },
                    "403": {
                        "description": "Forbidden"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "RoleRequest": {
                "title": "Role Form Request",
                "description": "role form request body data Validate",
                "required": [
                    "name",
                    "slug"
                ],
                "properties": {
                    "name": {
                        "title": "Role Name",
                        "description": "Name of the new Role",
                        "type": "string",
                        "example": "Hr Manager"
                    },
                    "slug": {
                        "title": "Role Slug",
                        "description": "Slug of the new Role",
                        "type": "string",
                        "example": "hr-manager"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RoleRequest"
                }
            },
...
            "RoleResource": {
                "title": "Role Resource",
                "description": "Role resource",
                "properties": {
                    "": {
                        "title": "Data",
                        "description": "Data wrapper",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "type": "object",
                "xml": {
                    "name": "RoleResource"
                }
            },
...
            "Role": {
                "title": "Role",
                "description": "Role model",
                "properties": {
                    "role_id": {
                        "title": "Role ID",
                        "description": "Role ID",
                        "type": "integer",
                        "format": "int64",
                        "example": 1
                    },
                    "name": {
                        "title": "Role Name",
                        "description": "Name of the new Role",
                        "type": "string",
                        "example": "Hr Manager"
                    },
                    "slug": {
                        "title": "Role Slug",
                        "description": "Slug of the new Role",
                        "type": "string",
                        "example": "hr-manager"
                    }
                },
                "type": "object",
                "xml": {
                    "name": "Role"
                }
            },
...
        }
    },
...
}